% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{marker_table<-}
\alias{marker_table<-}
\alias{assign-marker_table}
\title{Assign marker_table to \code{object}}
\usage{
marker_table(object) <- value
}
\arguments{
\item{object}{a \code{\linkS4class{microbiomeMarker}} object to modify.}

\item{value}{new value to replace the \code{marker_table} slot of \code{object}.
Either a \code{marker_table-class}, a \code{data.frame} that can be coerced
into \code{marker_table-class}.}
}
\description{
This function replace the \code{marker_table} slot of \code{object} with \code{value}.
}
