% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment-methods.R
\docType{methods}
\name{assign-otu_table}
\alias{assign-otu_table}
\alias{otu_table<-,microbiomeMarker,otu_table-method}
\alias{otu_table<-,microbiomeMarker,phyloseq-method}
\alias{otu_table<-,microbiomeMarker,microbiomeMarker-method}
\title{Assign a new OTU table}
\usage{
\S4method{otu_table}{microbiomeMarker,otu_table}(x) <- value

\S4method{otu_table}{microbiomeMarker,phyloseq}(x) <- value

\S4method{otu_table}{microbiomeMarker,microbiomeMarker}(x) <- value
}
\arguments{
\item{x}{\code{\linkS4class{microbiomeMarker}}}

\item{value}{\code{\linkS4class{otu_table}}, \code{\linkS4class{phyloseq}},
or \code{\linkS4class{microbiomeMarker}}}
}
\value{
a \code{\linkS4class{microbiomeMarker}} object.
}
\description{
Assign a new OTU table in microbiomeMarker object
}
