% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data-caporaso}
\alias{data-caporaso}
\alias{caporaso}
\title{16S rRNA data from "Moving pictures of the human microbiome"}
\format{
a \link[phyloseq:phyloseq]{phyloseq::phyloseq} object
}
\source{
Data was downloaded from https://www.microbiomeanalyst.ca
}
\description{
16S read counts and phylogenetic tree file of 34 Illumina samples derived
from Moving Pictures of the Human Microbiome (Caporaso et al.) Group label:
gut, left palm, right palm, and tongue - indicating different sampled body
sites.
}
\references{
Caporaso, et al. Moving pictures of the human microbiome. Genome Biol 12,
R50 (2011).

\url{https://doi.org/10.1186/gb-2011-12-5-r50}
}
\author{
Yang Cao
}
