% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data-cid_ying}
\alias{data-cid_ying}
\alias{cid_ying}
\title{16S rRNA data of 94 patients from CID 2012}
\format{
a \link[phyloseq:phyloseq]{phyloseq::phyloseq} object
}
\source{
\url{https://github.com/ying14/yingtools2/tree/master/data}
}
\description{
Data from a cohort of 94 Bone Marrow Transplant patients previously published
on in CID
}
\references{
Ying, et al. Intestinal Domination and the Risk of Bacteremia in Patients
Undergoing Allogeneic Hematopoietic Stem Cell Transplantation,
Clinical Infectious Diseases, Volume 55, Issue 7, 1 October 2012,
Pages 905–914,

\url{https://academic.oup.com/cid/article/55/7/905/428203}
}
\author{
Yang Cao
}
