% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data-ecam}
\alias{data-ecam}
\alias{ecam}
\title{Data from Early Childhood Antibiotics and the Microbiome (ECAM) study}
\format{
a \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}} object
}
\description{
The data from a subset of the Early Childhood Antibiotics and the
Microbiome (ECAM) study, which tracked the microbiome composition and
development of 43 infants in the United States from birth to 2 years of age,
identifying microbiome associations with antibiotic exposure, delivery mode,
and diet.
}
\references{
Bokulich, Nicholas A., et al. "Antibiotics, birth mode, and diet shape
microbiome maturation during early life." Science translational medicine
8.343 (2016): 343ra82-343ra82.

\url{https://github.com/FrederickHuangLin/ANCOM/tree/master/data}
}
