% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data-enterotypes_arumugam}
\alias{data-enterotypes_arumugam}
\alias{enterotypes_arumugam}
\title{Enterotypes data of 39 samples}
\format{
a \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}} object
}
\description{
The data contains 22 European metagenomes from Danish, French, Italian,
and Spanish individuals, and 13 Japanese and 4 American.
}
\references{
Arumugam, Manimozhiyan, et al. Enterotypes of the human gut microbiome.
nature 473.7346 (2011): 174-180.
}
\author{
Yang Cao
}
