% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data-kostic_crc}
\alias{data-kostic_crc}
\alias{kostic_crc}
\title{Data from a study on colorectal cancer (kostic 2012)}
\format{
a \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}} object
}
\description{
The data from a study on colorectal cancer. Samples that had no \code{DIAGNOSIS}
attribute assigned and with less than 500 reads (counts) were removed, and
191 samples remains (91 healthy and 86 Tumors).
}
\references{
Kostic et al. Genomic analysis identifies association of Fusobacterium with
colorectal carcinoma. Genome research, 2012, 22(2), 292-298.
}
\author{
Yang Cao
}
