% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data-oxygen}
\alias{data-oxygen}
\alias{oxygen}
\title{Oxygen availability 16S dataset, of which taxa table has been summarized for
python lefse input}
\format{
a \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}} object
}
\source{
\url{http://huttenhower.sph.harvard.edu/webfm_send/129}
}
\description{
A small subset of the HMP 16S dataset for finding biomarkers characterizing
different level of oxygen availability in different bodysites
}
\author{
Yang Cao
}
