% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data-pediatric_ibd}
\alias{data-pediatric_ibd}
\alias{pediatric_ibd}
\title{IBD stool samples}
\format{
a \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}} object
}
\source{
\url{https://www.microbiomeanalyst.ca/MicrobiomeAnalyst/resources}
}
\description{
43 pediatric IBD stool samples obtained from the Integrative Human Microbiome
Project Consortium (iHMP). Group label: CD and Controls.
}
\author{
Yang Cao
}
