% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data-spontaneous_colitis}
\alias{data-spontaneous_colitis}
\alias{spontaneous_colitis}
\title{This is a sample data from lefse python script, a 16S dataset for studying
the characteristics of the fecal microbiota in a mouse model of spontaneous
colitis.}
\format{
a \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}} object
}
\source{
\url{http://www.huttenhower.org/webfm_send/73}
}
\description{
The dataset contains 30 abundance profiles (obtained processing the 16S reads
with RDP) belonging to 10 rag2 (control) and 20 truc (case) mice.
}
\author{
Yang Cao
}
