% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-effect-size.R
\name{plot_ef_bar}
\alias{plot_ef_bar}
\alias{ef-barplot,ef-dotplot}
\alias{plot_ef_dot}
\title{bar and dot plot of effect size of microbiomeMarker data}
\usage{
plot_ef_bar(mm, label_level = 1, max_label_len = 60, n = 10)

plot_ef_dot(mm, label_level = 1, max_label_len = 60, n = 10)
}
\arguments{
\item{mm}{a \linkS4class{microbiomeMarker} object}

\item{label_level}{integer, number of label levels to be displayed, default
\code{1}, \code{0} means display the full name of the feature}

\item{max_label_len}{integer, maximum number of characters of feature label,
default \code{60}}

\item{n}{integer, number of markers to display, default \code{10}. If the number
of markers is less than 10, all markers will be displayed.}
}
\value{
a ggplot project
}
\description{
bar and dot plot of effect size microbiomeMarker data. This function returns
a \code{ggplot2} object that can be saved or further customized using \strong{ggplot2}
package.
}
