% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-hoc-test.R
\name{extract_posthoc_res}
\alias{extract_posthoc_res}
\title{Extract results from a posthoc test}
\usage{
extract_posthoc_res(object, features = NULL)
}
\arguments{
\item{object}{a \code{\linkS4class{postHocTest}} object.}

\item{features}{either \code{NULL} extracts results of all features, or a
character vector to specify the test resuts of which features are
extracted.}
}
\value{
a \code{\link[IRanges:DataFrameList-class]{IRanges::SimpleDFrameList}} object.
}
\description{
This function extracts the results of posthoc test.
}
\examples{
require(IRanges)
pht <- postHocTest(
    result = DataFrameList(
        featureA = DataFrame(
            comparisons = c("group2-group1",
                "group3-group1",
                "group3-group2"),
            diff_mean = runif(3),
            pvalue = rep(0.01, 3),
            ci_lower = rep(0.01, 3),
            ci_upper = rep(0.011, 3)
        ),
        featureB = DataFrame(
            comparisons = c("group2-group1",
                "group3-group1",
                "group3-group2"),
            diff_mean = runif(3),
            pvalue = rep(0.01, 3),
            ci_lower = rep(0.01, 3),
            ci_upper = rep(0.011, 3)
        )
    ),
    abundance = data.frame(
        featureA = runif(3),
        featureB = runif(3),
        group = c("group1", "group2", "grou3")
    )
)
extract_posthoc_res(pht, "featureA")[[1]]
}
