% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lefse_utilities.R
\name{fix_duplicate_tax}
\alias{fix_duplicate_tax}
\title{Duplicated taxa: e.g. maybe multiple species (s__uncultured)
belong to different genera. append the upper level taxa to the taxa to
distinguish this duplicated taxa}
\usage{
fix_duplicate_tax(ps)
}
\arguments{
\item{ps}{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq} object or \link[phyloseq:taxonomyTable-class]{phyloseq::taxonomyTable}
object}
}
\description{
Duplicated taxa: e.g. maybe multiple species (s__uncultured)
belong to different genera. append the upper level taxa to the taxa to
distinguish this duplicated taxa
}
\references{
https://github.com/lch14forever/microbiomeViz/blob/94cbfe452a735aadf88733b27b8221a03f450a55/R/utils.R#L68-L86
}
\keyword{internal}
