% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{get_norm_factors}
\alias{get_norm_factors}
\title{Extract the normalization factors}
\usage{
get_norm_factors(object)
}
\arguments{
\item{object}{a \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}}, \link[phyloseq:otu_table-class]{phyloseq::otu_table}
object.}
}
\value{
a numeric vector with the length equal to the number of samples, or
\code{NULL} if the \code{object} has not been normalized.
}
\description{
This function will be used to extract the normalization factors. After
dividing the observed feature table by normalization factors (eliminate
sequencing biases), we will obtain the normalized feature table.
}
