% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-cladogram.R
\name{get_treedata_phyloseq}
\alias{get_treedata_phyloseq}
\title{Generate tree data from phyloseq object}
\usage{
get_treedata_phyloseq(ps, sep = "|")
}
\arguments{
\item{ps}{a \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}} object}

\item{sep}{character, separate between different levels of taxa, default \code{|}}
}
\value{
a \code{\link[tidytree:treedata-class]{tidytree::treedata}} object
}
\description{
Generate tree data from phyloseq object
}
\author{
Yang Cao
}
\keyword{internal}
