% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lefse-cladogram.R
\name{get_unique_id}
\alias{get_unique_id}
\title{Get unique id for short label annotation}
\usage{
get_unique_id(n, depth = 1)
}
\description{
Get unique id for short label annotation
}
\references{
https://stackoverflow.com/questions/21681785/repeating-vector-of-letters/21689613#21689613
}
\keyword{internal}
