% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-qiime2.R
\name{import_qiime2}
\alias{import_qiime2}
\title{Import function to read the the output of dada2 as phyloseq object}
\usage{
import_qiime2(
  otu_qza,
  taxa_qza = NULL,
  sam_tab = NULL,
  refseq_qza = NULL,
  tree_qza = NULL
)
}
\arguments{
\item{otu_qza}{character, file path of the feature table from qiime2.}

\item{taxa_qza}{character, file path of the taxonomic table from qiime2,
default \code{NULL}.}

\item{sam_tab}{character, file path of the sample metadata in tsv format,
default \code{NULL}.}

\item{refseq_qza}{character, file path of the representative sequences from
qiime2, default \code{NULL}.}

\item{tree_qza}{character, file path of the phylogenetic tree from
qiime2, default \code{NULL}.}
}
\value{
\code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}} object.
}
\description{
Import the qiime2 artifacts, including feature table, taxonomic table,
phylogenetic tree, representative sequence and sample metadata into
phyloseq object.
}
\examples{
otuqza_file <- system.file(
    "extdata", "table.qza",
    package = "microbiomeMarker"
)
taxaqza_file <- system.file(
    "extdata", "taxonomy.qza",
    package = "microbiomeMarker"
)
sample_file <- system.file(
    "extdata", "sample-metadata.tsv",
    package = "microbiomeMarker"
)
treeqza_file <- system.file(
    "extdata", "tree.qza",
    package = "microbiomeMarker"
)
ps <- import_qiime2(
    otu_qza = otuqza_file, taxa_qza = taxaqza_file,
    sam_tab = sample_file, tree_qza = treeqza_file
)
ps
}
