% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_qiime2.R
\name{is_dna_seq}
\alias{is_dna_seq}
\title{check the row.names of feature table is DNA sequence or not
This function is from https://github.com/YuLab-SMU/MicrobiotaProcess/blob/master/R/import_qiime2.R#L169-L177}
\usage{
is_dna_seq(names)
}
\description{
check the row.names of feature table is DNA sequence or not
This function is from https://github.com/YuLab-SMU/MicrobiotaProcess/blob/master/R/import_qiime2.R#L169-L177
}
\keyword{internal}
