% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{keep_taxa_in_rows}
\alias{keep_taxa_in_rows}
\title{Transpose the phyloseq object to ensure taxa are in rows}
\usage{
keep_taxa_in_rows(ps)
}
\arguments{
\item{ps}{a \link[phyloseq:phyloseq-class]{phyloseq::phyloseq} object}
}
\description{
Transpose the phyloseq object to ensure taxa are in rows
}
\keyword{internal}
