% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lefse_visualization.R
\name{lefse_barplot}
\alias{lefse_barplot}
\title{lefse bar plot}
\usage{
lefse_barplot(lefse_out, lda_threshold)
}
\arguments{
\item{lefse_out}{data frame, out of \code{\link{lefse}}}

\item{lda_threshold}{threshold of lda score, features whose lda score is less
then it will not be plotted}
}
\value{
a ggplot project
}
\description{
lefse bar plot
}
