% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lefse-cladogram.R
\name{lefse_cladogram}
\alias{lefse_cladogram}
\title{plot cladogram of lefse results}
\usage{
lefse_cladogram(
  mm,
  color = (scales::hue_pal())(length(unique(mm@marker_table$enrich_group))),
  branch_size = 0.2,
  alpha = 0.2,
  node_size_scale = 1,
  node_size_offset = 1,
  clade_label_level = 4
)
}
\arguments{
\item{mm}{a \linkS4class{microbiomeMarker} object}

\item{color}{a color vector, used to highlight the clades of micribome
biomaker}

\item{branch_size}{numberic, size of branch, default \code{0.2}}

\item{alpha}{alpha parameter for shading, default \code{0.2}}

\item{node_size_scale}{the parameter 'a' controlling node size:
\code{node_size=a*log(relative_abundance) + b}}

\item{node_size_offset}{the parameter 'b' controlling node size:
\code{node_size=a*log(relative_abundance) + b}}

\item{clade_label_level}{max level of taxa used to label the clade, other
level of taxa  will be shown on the side}
}
\value{
a ggtree object
}
\description{
annotate a ggtree plot to highlight certain clades
}
\references{
This function is modified from \code{clada.anno} from microbiomeViz.
\url{https://github.com/lch14forever/microbiomeViz/blob/master/R/visualizer.R}
}
\seealso{
\code{\link[ggtree:ggtree]{ggtree::ggtree()}}
}
\author{
Chenhao Li, Guangchuang Yu, Chenghao Zhu, Yang Cao
}
