% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{marker_table}
\alias{marker_table}
\alias{marker_table,data.frame-method}
\alias{marker_table,}
\alias{data.frame-method}
\title{Build or access the marker_table}
\usage{
marker_table(object)

\S4method{marker_table}{data.frame}(object)
}
\arguments{
\item{object}{an object among the set of classes difined by the
microbiomeMarker package that contain \code{marker_table}}
}
\description{
This is the recommended function for both building and accessing microbiome
marker table (\code{\linkS4class{marker_table}}).
}
