% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{microbiomeMarker-class}
\alias{microbiomeMarker-class}
\alias{.microbiomeMarker}
\title{The main class for microbiomeMarker data}
\description{
\code{microbiomeMarker-class} is inherited from the \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}}
by adding a custom slot \code{microbiome_marker} to save the differential analysis
results. And it provides a seamless interface with \strong{phyloseq}, which makes
\strong{microbiomeMarker} simple and easy to use. For more details on see the
document of \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}}.
}
\section{Slots}{

\describe{
\item{\code{microbiome_marker}}{a class of data.frame, the result of microbiome
differntial analysis
\itemize{
\item \code{lda}, logarithmic LDA score
\item \code{p_value}, p value of kw test
}}
}}

\seealso{
\code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}}
}
