% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/microbiomeMarker.R
\docType{package}
\name{microbiomeMarker}
\alias{microbiomeMarker}
\alias{microbiomeMarker-package}
\title{Build microbiomeMarker-class objects from their slots}
\usage{
microbiomeMarker(microbiome_marker, ...)
}
\arguments{
\item{microbiome_marker}{a data frame, the result of microbiome
differtial analysis}

\item{...}{arguments passed to \code{\link[phyloseq:phyloseq]{phyloseq::phyloseq()}}}
}
\description{
This the constructor to build the \code{microbiomeMarker-class} object, don't use
the \code{new()} constructor.

The microboimeMarker package provides several methods to identify micribome
biomarker, such as lefse and statistical analysis.
}
\references{
\href{https://stackoverflow.com/questions/9900134/is-it-bad-practice-to-access-s4-objects-slots-directly-using/9900822#9900822}{Is it bad practice to access S4 objects slots directly using @?}
}
\seealso{
\code{\link[phyloseq:phyloseq]{phyloseq::phyloseq()}},
}
