% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{nclass}
\alias{nclass}
\alias{nclass,microbiomeMarker-method}
\alias{nclass,marker_table-method}
\title{Get the number of classes for microbiome discovery}
\usage{
nclass(object)

\S4method{nclass}{microbiomeMarker}(object)

\S4method{nclass}{marker_table}(object)
}
\arguments{
\item{object}{a \code{\linkS4class{microbiomeMarker}} or \code{\linkS4class{marker_table}} object}
}
\value{
an integer
}
\description{
Get the number of classes for microbiome discovery
}
