% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{nmarker}
\alias{nmarker}
\alias{nmarker,microbiomeMarker-method}
\alias{nmarker,marker_table-method}
\title{Get the number of microbiome markers}
\usage{
nmarker(object)

\S4method{nmarker}{microbiomeMarker}(object)

\S4method{nmarker}{marker_table}(object)
}
\arguments{
\item{object}{a \code{\linkS4class{microbiomeMarker}} or \code{\linkS4class{marker_table}} object}
}
\value{
an integer, the number of microbiome markers
}
\description{
Get the number of microbiome markers
}
