% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{norm_clr}
\alias{norm_clr}
\title{CLR (centered log-ratio) normalization}
\usage{
norm_clr(object)
}
\arguments{
\item{object}{a \link[phyloseq:phyloseq-class]{phyloseq::phyloseq} or \link[phyloseq:otu_table-class]{phyloseq::otu_table}
object}
}
\description{
CLR (centered log-ratio) normalization
}
\keyword{internal}
