% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{norm_css}
\alias{norm_css}
\title{css normalization (metagenomeSeq)}
\usage{
norm_css(object, log = FALSE, sl = 1000)
}
\arguments{
\item{object}{a \link[phyloseq:phyloseq-class]{phyloseq::phyloseq} or \link[phyloseq:otu_table-class]{phyloseq::otu_table}
object}

\item{log}{logical, whether or not to log2 transfrom scale,
see \code{\link[metagenomeSeq:MRcounts]{metagenomeSeq::MRcounts()}}.}

\item{sl}{The value to scale, see \code{\link[metagenomeSeq:MRcounts]{metagenomeSeq::MRcounts()}}}
}
\description{
css normalization (metagenomeSeq)
}
\seealso{
\code{\link[metagenomeSeq:cumNorm]{metagenomeSeq::cumNorm()}}, \code{\link[metagenomeSeq:MRcounts]{metagenomeSeq::MRcounts()}}
}
\keyword{internal}
