% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{norm_rarefy}
\alias{norm_rarefy}
\title{rarefying}
\usage{
norm_rarefy(
  object,
  size = min(sample_sums(object)),
  rng_seed = FALSE,
  replace = TRUE,
  trim_otus = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{object}{a \link[phyloseq:phyloseq-class]{phyloseq::phyloseq} or \link[phyloseq:otu_table-class]{phyloseq::otu_table} object}

\item{size, rng_seed, replace, trim_otus, verbose}{for details see
\code{\link[phyloseq:rarefy_even_depth]{phyloseq::rarefy_even_depth()}}}
}
\description{
rarefying
}
\keyword{internal}
