% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{norm_rle}
\alias{norm_rle}
\title{Relative log expression (RLE) normalization}
\usage{
norm_rle(
  object,
  locfunc = stats::median,
  geo_means = NULL,
  control_genes = NULL
)
}
\arguments{
\item{object}{a \link[phyloseq:phyloseq-class]{phyloseq::phyloseq} or \link[phyloseq:otu_table-class]{phyloseq::otu_table}
object}

\item{geo_means}{default \code{NULL}, which means the geometric means of the
counts are used. A vector of geometric means from another count matrix can
be provided for a "frozen" size factor calculation.}

\item{control_genes}{default \code{NULL}, which means all genes are used for size
factor estimation, numeric or logical index vector specifying the genes
used for size factor estimation (e.g. housekeeping or spike-in genes).}

\item{logcfunc}{a function to compute a location for a sample. By default,
the median is used.}
}
\description{
Relative log expression (RLE) normalization
}
\seealso{
\code{\link[DESeq2:estimateSizeFactorsForMatrix]{DESeq2::estimateSizeFactorsForMatrix()}}
}
\keyword{internal}
