% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{norm_tmm}
\alias{norm_tmm}
\title{TMM (trimmed mean of m-values) normalization}
\usage{
norm_tmm(
  object,
  ref_column = NULL,
  logratio_trim = 0.3,
  sum_trim = 0.05,
  do_weighting = TRUE,
  Acutoff = -1e+10
)
}
\arguments{
\item{object}{a \link[phyloseq:phyloseq-class]{phyloseq::phyloseq} or \link[phyloseq:otu_table-class]{phyloseq::otu_table}
object}

\item{ref_column}{column to use as reference}

\item{logratio_trim}{amount of trim to use on log-ratios}

\item{sum_trim}{amount of trim to use on the combined absolute levels
("A" values)}

\item{do_weighting}{whether to compute the weights or not}

\item{Acutoff}{cutoff on "A" values to use before trimming}
}
\description{
TMM (trimmed mean of m-values) normalization
}
\references{
https://github.com/biobakery/Maaslin2/blob/master/R/utility_scripts.R
}
\seealso{
\code{\link[edgeR:calcNormFactors]{edgeR::calcNormFactors()}}
}
\keyword{internal}
