% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{norm_tss}
\alias{norm_tss}
\title{TSS normalization (relative abundance)}
\usage{
norm_tss(object)
}
\arguments{
\item{object}{object a \link[phyloseq:phyloseq-class]{phyloseq::phyloseq} or
\link[phyloseq:otu_table-class]{phyloseq::otu_table} object}
}
\description{
TSS normalization (relative abundance)
}
\keyword{internal}
