% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{norm_value}
\alias{norm_value}
\title{Normalize the sum of values of each sample to a given value}
\usage{
norm_value(object, normalization)
}
\arguments{
\item{object}{a \link[phyloseq:phyloseq-class]{phyloseq::phyloseq} or \link[phyloseq:otu_table-class]{phyloseq::otu_table}}

\item{normalization}{set the normalization value, return the feature itself
if not supported.}
}
\description{
Normalize the sum of values of each sample to a given value
}
\keyword{internal}
