% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalize,phyloseq-method}
\alias{normalize,phyloseq-method}
\alias{normalize,otu_table-method}
\alias{normalize}
\alias{normalize,data.frame-method}
\alias{normalize,matrix-method}
\title{Normalize the microbial abundance data}
\usage{
\S4method{normalize}{phyloseq}(object, method = "TSS", ...)

\S4method{normalize}{otu_table}(object, method = "TSS", ...)

\S4method{normalize}{data.frame}(object, method = "TSS", ...)

\S4method{normalize}{matrix}(object, method = "TSS", ...)
}
\arguments{
\item{object}{a matrix, data.frame, \link[phyloseq:phyloseq-class]{phyloseq::phyloseq} or
\link[phyloseq:otu_table-class]{phyloseq::otu_table} object}

\item{method}{the methods used to normalize the microbial abundance data.
Options includes:
\itemize{
\item a integer, e.g. 1e6, indicating pre-sample normalization of the sum of
the values to 1e6.
\item "none": do not normalize.
\item "rarefy": random subsampling counts to the smallest library size in the
data set.
\item "TSS": total sum scaling, also referred to as "relative abundance", the
abundances were normalized by dividing the corresponding sample library
size.
\item "TMM": trimmed mean of m-values. First, a sample is chosen as reference.
The scaling factor is then derived using a weighted trimmed mean over the
differences of the log-transformed gene-count fold-change between the
sample and the reference.
\item "RLE", relative log expression, RLE uses a pseudo-reference calculated
using the geometric mean of the gene-specific abundances over all
samples. The scaling factors are then calculated as the median of the
gene counts ratios between the samples and the reference.
\item "CSS": cumulative sum scaling, calculates scaling factors as the
cumulative sum of gene abundances up to a data-derived threshold.
\item "CLR": centered log-ratio normalization.
}}

\item{...}{other arguments passed to the corresponding normalization methods.}
}
\description{
Normalize the microbial abundance data
}
\seealso{
\code{\link[edgeR:calcNormFactors]{edgeR::calcNormFactors()}},\code{\link[DESeq2:estimateSizeFactorsForMatrix]{DESeq2::estimateSizeFactorsForMatrix()}},
\code{\link[metagenomeSeq:cumNorm]{metagenomeSeq::cumNorm()}}
}
