% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_qiime2.R
\name{parse_q2taxonomy}
\alias{parse_q2taxonomy}
\title{Parse qiime2 taxa in different taxonomic levels}
\usage{
parse_q2taxonomy(taxa, sep = "; |;", trim_rank_prefix = TRUE)
}
\arguments{
\item{taxa}{\code{tax_table} object.}

\item{sep}{character string containing a regular expression, separator
between different taxonomic levels, defaults to on compatible with both
GreenGenes and SILVA \verb{; |;"}.}

\item{trim_rank_prefix}{logical whether remove leading characters from
taxonomic levels, e.g. k__ or D_0__, default \code{TRUE}.}
}
\value{
\code{\link[phyloseq:taxonomyTable-class]{phyloseq::taxonomyTable}} object.
}
\description{
Parse qiime2 taxa in different taxonomic levels
}
\keyword{internal}
