% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA-deseq2.R
\name{phyloseq2DESeq2}
\alias{phyloseq2DESeq2}
\title{Convert \code{phyloseq-class} object to \code{DESeqDataSet-class} object}
\usage{
phyloseq2DESeq2(ps, design, ...)
}
\arguments{
\item{ps}{the [phyloseq::phyloseq-class\verb{] object to convert, which must have a [}phyloseq::sample_data()`] component.}

\item{design}{a \code{formula} or \code{matrix}, the formula expresses how the counts
for each gene depend on the variables in colData. Many R formula are valid,
including designs with multiple variables, e.g., \code{~ group + condition}.
This argument is passed to \code{\link[DESeq2:DESeqDataSet]{DESeq2::DESeqDataSetFromMatrix()}}.}

\item{...}{additional arguments passed to
\code{\link[DESeq2:DESeqDataSet]{DESeq2::DESeqDataSetFromMatrix()}}, Most users will not need to pass any
additional arguments here.}
}
\value{
a \code{\link[DESeq2:DESeqDataSet]{DESeq2::DESeqDataSet}} object.
}
\description{
This function convert [phyloseq::phyloseq-class\verb{] to [}DESeq2::DESeqDataSet-class\verb{], which can then be tested using [}DESeq2::DESeq()`].
}
\examples{
data(caporaso)
phyloseq2DESeq2(caporaso, ~SampleType)
}
\seealso{
\code{\link[DESeq2:DESeqDataSet]{DESeq2::DESeqDataSetFromMatrix()}},\code{\link[DESeq2:DESeq]{DESeq2::DESeq()}}
}
