% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA-edgeR.R
\name{phyloseq2edgeR}
\alias{phyloseq2edgeR}
\title{Convert phyloseq data to edgeR \code{DGEList} object}
\usage{
phyloseq2edgeR(ps, ...)
}
\arguments{
\item{ps}{a \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}} object.}

\item{...}{optional, additional named arguments passed  to
\code{\link[edgeR:DGEList]{edgeR::DGEList()}}. Most users will not need to pass any additional
arguments here.}
}
\value{
A \code{\link[edgeR:DGEList-class]{edgeR::DGEList}} object.
}
\description{
This function convert \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}} object to
\code{\link[edgeR:DGEList-class]{edgeR::DGEList}} object, can then can be used to perform
differential analysis using the methods in \strong{edgeR}.
}
\examples{
data(caporaso)
dge <- phyloseq2edgeR(caporaso)
}
