% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-comparing.R
\name{plot.compareDA}
\alias{plot.compareDA}
\title{Plotting DA comparing result}
\usage{
\method{plot}{compareDA}(x, sort = c("score", "auc", "fpr", "power"), ...)
}
\arguments{
\item{x}{an \code{compareDA} object, output from \code{\link[=compare_DA]{compare_DA()}}.}

\item{sort}{character string specifying sort method. Possibilities are
"score" which is calculated as \eqn{(auc - 0.5) * power - fdr}, "auc" for
area under the ROC curve, "fpr" for false positive rate, "power" for
empirical power.}

\item{...}{extra arguments, just ignore it.}
}
\value{
a \code{\link[ggplot2:ggplot]{ggplot2::ggplot}} object containing 4 subplots: "auc", "fdr",
"power"and "score" plot.
}
\description{
Plotting DA comparing result
}
