% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-abundance.R
\name{plot_abundance}
\alias{plot_abundance}
\title{plot the abundances of markers}
\usage{
plot_abundance(mm, label_level = 1, max_label_len = 60, markers = NULL, group)
}
\arguments{
\item{mm}{a \code{\linkS4class{microbiomeMarker}} object}

\item{label_level}{integer, number of label levels to be displayed, default
\code{1}, \code{0} means display the full name of the feature}

\item{max_label_len}{integer, maximum number of characters of feature label,
default \code{60}}

\item{markers}{character vector, markers to display, default \code{NULL},
indicating plot all markers.}

\item{group}{character, the variable to set the group}
}
\value{
a \code{\link[ggplot2:ggplot]{ggplot2::ggplot}} object.
}
\description{
plot the abundances of markers
}
\examples{
data(enterotypes_arumugam)
mm <- run_limma_voom(
    enterotypes_arumugam,
    "Enterotype",
    contrast = c("Enterotype 3", "Enterotype 2"),
    pvalue_cutoff = 0.01,
    p_adjust = "none"
)
plot_abundance(mm, group = "Enterotype")
}
