% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lefse-cladogram.R
\name{plot_cladogram}
\alias{plot_cladogram}
\title{plot cladogram of micobiomeMaker results}
\usage{
plot_cladogram(
  mm,
  color,
  branch_size = 0.2,
  alpha = 0.2,
  node_size_scale = 1,
  node_size_offset = 1,
  clade_label_level = 4,
  annotation_shape = 22,
  annotation_shape_size = 5,
  group_legend_param = list(),
  marker_legend_param = list()
)
}
\arguments{
\item{mm}{a \linkS4class{microbiomeMarker} object}

\item{color}{a color vector, used to highlight the clades of microbiome
biomarker. The values will be matched in order (usually alphabetical) with
the groups. If this is a named vector, then the colors will be matched
based on the names instead.}

\item{branch_size}{numberic, size of branch, default \code{0.2}}

\item{alpha}{alpha parameter for shading, default \code{0.2}}

\item{node_size_scale}{the parameter 'a' controlling node size:
\code{node_size=a*log(relative_abundance) + b}}

\item{node_size_offset}{the parameter 'b' controlling node size:
\code{node_size=a*log(relative_abundance) + b}}

\item{clade_label_level}{max level of taxa used to label the clade, other
level of taxa  will be shown on the side}

\item{annotation_shape}{shape used for annotation, default \code{22}}

\item{annotation_shape_size}{size used for annotation shape, default \code{5}}

\item{group_legend_param, marker_legend_param}{a list specifying
extra parameters of group legend and marker legend, such as \code{direction} (
the direction of the guide), \code{nrow} (the desired number of rows of
legends). See \code{\link[ggplot2:guide_legend]{ggplot2::guide_legend()}} for more details.}
}
\value{
a ggtree object
}
\description{
plot cladogram of micobiomeMaker results
}
\references{
This function is modified from \code{clada.anno} from microbiomeViz.
\url{https://github.com/lch14forever/microbiomeViz/blob/master/R/visualizer.R}
}
\seealso{
\code{\link[ggtree:ggtree]{ggtree::ggtree()}}, \code{\link[=set_marker_annotation]{set_marker_annotation()}}
}
\author{
Chenhao Li, Guangchuang Yu, Chenghao Zhu, Yang Cao
}
