% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-bar.R
\name{plot_ef_bar}
\alias{plot_ef_bar}
\title{barplot of effect size of microbiomeMarker data}
\usage{
plot_ef_bar(mm, label_level = 1, max_label_len = 60, direction = c("h", "v"))
}
\arguments{
\item{mm}{a \linkS4class{microbiomeMarker} object}

\item{label_level}{integer, number of label levels to be displayed, default
\code{1}, \code{0} means display the full name of the feature}

\item{max_label_len}{integer, maximum number of characters of feature labels,
default \code{60}}

\item{direction}{the direction of bar, horizontal (\code{h}) or vertical (\code{v}),
default \code{h}}
}
\value{
a ggplot project
}
\description{
barplot of effect size microbiomeMarker data. This function returns a
\code{ggplot2} object that can be saved or further customized using \strong{ggplot2}.
}
