% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Heatmap of microbiome marker}
\usage{
plot_heatmap(
  mm,
  transform = c("log10", "log10p", "identity"),
  cluster_marker = FALSE,
  cluster_sample = FALSE,
  markers = NULL,
  label_level = 1,
  max_label_len = 60,
  sample_label = FALSE,
  scale_by_row = FALSE,
  annotation_col = NULL,
  group,
  ...
)
}
\arguments{
\item{mm}{a \code{\linkS4class{microbiomeMarker}} object}

\item{transform}{transformation to apply, for more details see
\code{\link[=transform_abundances]{transform_abundances()}}:
\itemize{
\item "identity", return the original data without any transformation.
\item "log10", the transformation is \code{log10(object)}, and if the data contains
zeros the transformation is \code{log10(1 + object)}.
\item "log10p", the transformation is \code{log10(1 + object)}.
}}

\item{cluster_marker, cluster_sample}{logical, controls whether to perform
clustering in markers (rows) and samples (cols), default \code{FALSE}.}

\item{markers}{character vector, markers to display, default \code{NULL},
indicating plot all markers.}

\item{label_level}{integer, number of label levels to be displayed, default
\code{1}, \code{0} means display the full name of the feature}

\item{max_label_len}{integer, maximum number of characters of feature label,
default \code{60}}

\item{sample_label}{logical, controls whether to show the sample labels in
the heatmap, default \code{FALSE}.}

\item{scale_by_row}{logical, controls whether to scale the heatmap by the
row (marker) values, default \code{FALSE}.}

\item{annotation_col}{assign colors for the top annotation using a named
vector, passed to \code{col} in \code{\link[ComplexHeatmap:HeatmapAnnotation]{ComplexHeatmap::HeatmapAnnotation()}}.}

\item{group}{character, the variable to set the group}

\item{...}{extra arguments passed to \code{\link[ComplexHeatmap:Heatmap]{ComplexHeatmap::Heatmap()}}.}
}
\value{
a \code{\link[ComplexHeatmap:Heatmap-class]{ComplexHeatmap::Heatmap}} object.
}
\description{
Display the microbiome marker using heatmap, in which rows represents the
marker and columns represents the samples.
}
\examples{
data(kostic_crc)
kostic_crc_small <- phyloseq::subset_taxa(
    kostic_crc,
    Phylum \%in\% c("Firmicutes")
)
mm_lefse <- run_lefse(
    kostic_crc_small,
    wilcoxon_cutoff = 0.01,
    group = "DIAGNOSIS",
    kw_cutoff = 0.01,
    multigrp_strat = TRUE,
    lda_cutoff = 4
)
plot_heatmap(mm_lefse, group = "DIAGNOSIS")
}
\seealso{
\code{\link{transform_abundances}},\code{\link[ComplexHeatmap:Heatmap]{ComplexHeatmap::Heatmap()}}
}
