% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-sl-roc.R
\name{plot_sl_roc}
\alias{plot_sl_roc}
\title{ROC curve of microbiome marker from supervised learning methods}
\usage{
plot_sl_roc(mm, group, nfolds = 3, nrepeats = 3, tune_length = 5, ...)
}
\arguments{
\item{mm}{a \linkS4class{microbiomeMarker} object.}

\item{group, nfolds, nrepeats, tune_length, ...}{same with the \code{run_sl()}.}
}
\value{
a \code{\link[ggplot2:ggplot]{ggplot2::ggplot}} object.
}
\description{
Show the ROC curve of the microbiome marker calculated by \code{run_sl}.
}
\seealso{
\code{\link[=run_sl]{run_sl()}}
}
