% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-multiple-groups.R
\name{posthoc_test}
\alias{posthoc_test}
\title{Post hoc pairwise comparisons for multiple groups test.}
\usage{
posthoc_test(
  ps,
  group,
  rank_name,
  conf_level = 0.95,
  method = c("tukey", "games_howell", "scheffe", "welch_uncorrected")
)
}
\arguments{
\item{ps}{a \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}} object}

\item{group}{character, the variable to set the group}

\item{rank_name}{character, taxonomic names of \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}}
to compare}

\item{conf_level}{confidence level, default 0.95}

\item{method}{one of "tukey", "games_howell", "scheffe", "welch_uncorrected",
defining the method for the pairwise comparisons. See details for more
information.}
}
\value{
a \linkS4class{postHocTest} object
}
\description{
Multiple group test, such as anova and Kruskal-Wallis rank sum test, can be
used to uncover the significant feature among all groups. Post hoc tests are
used to uncover specific mean differences between pair of groups.
}
\seealso{
\linkS4class{postHocTest}, \code{\link[=test_multiple_groups]{test_multiple_groups()}}
}
