% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllGenerics.R
\docType{class}
\name{postHocTest-class}
\alias{postHocTest-class}
\alias{show,postHocTest-method}
\alias{show,}
\alias{postHocTest-method}
\title{The postHocTest Class, represents the result of post-hoc test result among
multiple groups}
\usage{
\S4method{show}{postHocTest}(object)
}
\arguments{
\item{object}{a \code{postHocTest-class} object}
}
\value{
a \code{\linkS4class{postHocTest}} object.
}
\description{
The postHocTest Class, represents the result of post-hoc test result among
multiple groups
}
\section{Slots}{

\describe{
\item{\code{result}}{a \code{\link[IRanges:DataFrameList-class]{IRanges::DataFrameList}}, each \code{DataFrame} consists
of five variables:
\itemize{
\item \code{comparisons}: character, specify which two groups to test (the group names
are separated by "_)
\item \code{diff_mean}: numeric, difference in mean abundances
\item \code{pvalue}: numeric, p values
\item \code{ci_lower} and \code{ci_upper}: numeric, lower and upper confidence interval of
difference in mean abundances
}}

\item{\code{abundance}}{abundance of each feature in each group}

\item{\code{conf_level}}{confidence level}

\item{\code{method}}{method used for post-hoc test}

\item{\code{method_str}}{method illustration}
}}

\author{
Yang Cao
}
