% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{postHocTest}
\alias{postHocTest}
\title{Build postHocTest object}
\usage{
postHocTest(
  result,
  abundance,
  conf_level = 0.95,
  method = "tukey",
  method_str = paste("Posthoc multiple comparisons of means: ", method)
)
}
\arguments{
\item{result}{a \code{\link[IRanges:DataFrameList-class]{IRanges::SimpleDFrameList}} object.}

\item{abundance}{data.frame.}

\item{conf_level}{numeric, confidence level.}

\item{method}{character, method for posthoc test.}

\item{method_str}{character, illustrates which method is used for posthoc
test.}
}
\value{
a \code{\linkS4class{postHocTest}} object.
}
\description{
This function is used for create \code{postHocTest} object, and is only used for
developers.
}
\examples{
require(IRanges)
pht <- postHocTest(
    result = DataFrameList(
        featureA = DataFrame(
            comparisons = c("group2-group1", 
                "group3-group1", 
                "group3-group2"),
            diff_mean = runif(3),
            pvalue = rep(0.01, 3),
            ci_lower = rep(0.01, 3),
            ci_upper = rep(0.011, 3)
        ),
        featureB = DataFrame(
            comparisons = c("group2-group1", 
                "group3-group1", 
                "group3-group2"),
            diff_mean = runif(3),
            pvalue = rep(0.01, 3),
            ci_lower = rep(0.01, 3),
            ci_upper = rep(0.011, 3)
        )
    ),
    abundance = data.frame(
        featureA = runif(3),
        featureB = runif(3),
        group = c("group1", "group2", "grou3")
    )
)
pht
}
