% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_qiime2.R
\name{read_q2biom}
\alias{read_q2biom}
\title{Read qiime2 feature table}
\usage{
read_q2biom(file)
}
\arguments{
\item{file}{character, file name of the biom file.}
}
\value{
\code{\link[phyloseq:otu_table-class]{phyloseq::otu_table}} object.
}
\description{
Import qiime2 feature table into otu_table object
}
