% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_qiime2.R
\name{read_qza}
\alias{read_qza}
\title{Read the qza file output from qiime2}
\usage{
read_qza(file, temp = tempdir())
}
\arguments{
\item{file}{character, path of the input qza file. Only files in format of
\code{BIOMV210DirFmt} (feature table), \code{TSVTaxonomyDirectoryFormat} (taxonomic
table), \code{NewickDirectoryFormat} (phylogenetic tree ) and
\code{DNASequencesDirectoryFormat} (representative sequences) are supported
right now.}

\item{temp}{character, a temporary directory in which the qza file will be
decompressed to, default \code{tempdir()}.}
}
\value{
\code{\link[phyloseq:otu_table-class]{phyloseq::otu_table}} object for feature table,
\code{\link[phyloseq:taxonomyTable-class]{phyloseq::taxonomyTable}} object for taxonomic table,
\code{\link[ape:read.tree]{ape::phylo}} object for phylogenetic tree,
\code{\link[Biostrings:XStringSet-class]{Biostrings::DNAStringSet}} for representative sequences of taxa.
}
\description{
Import the qiime2 artifacts to R.
}
