% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA-aldex.R
\name{run_aldex}
\alias{run_aldex}
\title{Perform differential analysis using ALDEx2}
\usage{
run_aldex(
  ps,
  group,
  taxa_rank = "all",
  transform = c("identity", "log10", "log10p"),
  norm = "none",
  norm_para = list(),
  method = c("t.test", "wilcox.test", "kruskal", "glm_anova"),
  p_adjust = c("none", "fdr", "bonferroni", "holm", "hochberg", "hommel", "BH", "BY"),
  pvalue_cutoff = 0.05,
  mc_samples = 128,
  denom = c("all", "iqlr", "zero", "lvha"),
  paired = FALSE
)
}
\arguments{
\item{ps}{a \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}} object}

\item{group}{character, the variable to set the group}

\item{taxa_rank}{character to specify taxonomic rank to perform
differential analysis on. Should be one of
\code{phyloseq::rank_names(phyloseq)}, or "all" means to summarize the taxa by
the top taxa ranks (\code{summarize_taxa(ps, level = rank_names(ps)[1])}), or
"none" means perform differential analysis on the original taxa
(\code{taxa_names(phyloseq)}, e.g., OTU or ASV).}

\item{transform}{character, the methods used to transform the microbial
abundance. See \code{\link[=transform_abundances]{transform_abundances()}} for more details. The
options include:
\itemize{
\item "identity", return the original data without any transformation
(default).
\item "log10", the transformation is \code{log10(object)}, and if the data contains
zeros the transformation is \code{log10(1 + object)}.
\item "log10p", the transformation is \code{log10(1 + object)}.
}}

\item{norm}{the methods used to normalize the microbial abundance data. See
\code{\link[=normalize]{normalize()}} for more details.
Options include:
\itemize{
\item "none": do not normalize.
\item "rarefy": random subsampling counts to the smallest library size in the
data set.
\item "TSS": total sum scaling, also referred to as "relative abundance", the
abundances were normalized by dividing the corresponding sample library
size.
\item "TMM": trimmed mean of m-values. First, a sample is chosen as reference.
The scaling factor is then derived using a weighted trimmed mean over the
differences of the log-transformed gene-count fold-change between the
sample and the reference.
\item "RLE", relative log expression, RLE uses a pseudo-reference calculated
using the geometric mean of the gene-specific abundances over all
samples. The scaling factors are then calculated as the median of the
gene counts ratios between the samples and the reference.
\item "CSS": cumulative sum scaling, calculates scaling factors as the
cumulative sum of gene abundances up to a data-derived threshold.
\item "CLR": centered log-ratio normalization.
\item "CPM": pre-sample normalization of the sum of the values to 1e+06.
}}

\item{norm_para}{arguments passed to specific normalization methods}

\item{method}{test method, options include: "t.test" and "wilcox.test"
for two groups comparison,  "kruskal" and "glm_anova" for multiple groups
comparison.}

\item{p_adjust}{method for multiple test correction, default \code{none},
for more details see \link[stats:p.adjust]{stats::p.adjust}.}

\item{pvalue_cutoff}{cutoff of p value, default 0.05.}

\item{mc_samples}{integer, the number of Monte Carlo samples to use for
underlying distributions estimation, 128 is usually sufficient.}

\item{denom}{character string, specifiy which features used to as the
denominator for the geometric mean calculation. Options are:
\itemize{
\item "all", with all features.
\item "iqlr", accounts for data with systematic variation and centers the
features on the set features that have variance that is between the lower
and upper quartile of variance.
\item "zero", a more extreme case where there are many non-zero features in
one condition but many zeros in another. In this case the geometric mean
of each group is calculated using the set of per-group non-zero features.
\item "lvha", with house keeping features.
}}

\item{paired}{logical, whether to perform paired tests, only worked for
method "t.test" and "wilcox.test".}
}
\value{
a \code{\linkS4class{microbiomeMarker}} object.
}
\description{
Perform differential analysis using ALDEx2
}
\examples{
data(enterotypes_arumugam)
ps <- phyloseq::subset_samples(
    enterotypes_arumugam,
    Enterotype \%in\% c("Enterotype 3", "Enterotype 2")
)
run_aldex(ps, group = "Enterotype")
}
\references{
Fernandes, A.D., Reid, J.N., Macklaim, J.M. et al. Unifying the
analysis of high-throughput sequencing datasets: characterizing RNA-seq,
16S rRNA gene sequencing and selective growth experiments by compositional
data analysis. Microbiome 2, 15 (2014).
}
\seealso{
\code{\link[ALDEx2:aldex]{ALDEx2::aldex()}}
}
