% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancom.R
\name{run_ancom}
\alias{run_ancom}
\title{Perform differential analysis using ANCOM}
\usage{
run_ancom(
  ps,
  group_var,
  transform = c("identity", "log10", "log10p"),
  norm = "TSS",
  norm_para = list(),
  p_adjust = c("none", "fdr", "bonferroni", "holm", "hochberg", "hommel", "BH", "BY"),
  alpha = 0.05,
  theta = 0.75,
  test = c("aov", "wilcox.test", "kruskal.test"),
  ...
)
}
\arguments{
\item{ps}{a \code{\link[phyloseq]{phyloseq-class}} object.}

\item{group_var}{character, the variable to set the group.}

\item{transform}{character, the methods used to transform the microbial
abundance. See \code{\link[=transform_abundances]{transform_abundances()}} for more details. The
options include:
\itemize{
\item "identity", return the original data without any transformation (default).
\item "log10", the transformation is \code{log10(object)}, and if the data contains
zeros the transformation is \code{log10(1 + object)}.
\item "log10p", the transformation is \code{log10(1 + object)}.
}}

\item{norm}{the methods used to normalize the microbial abundance data. See
\code{\link[=normalize]{normalize()}} for more details.
Options include:
\itemize{
\item a integer, e.g. 1e6 (default), indicating pre-sample normalization of
the sum of the values to 1e6.
\item "none": do not normalize.
\item "rarefy": random subsampling counts to the smallest library size in the
data set.
\item "TSS": total sum scaling, also referred to as "relative abundance", the
abundances were normalized by dividing the corresponding sample library
size.
\item "TMM": trimmed mean of m-values. First, a sample is chosen as reference.
The scaling factor is then derived using a weighted trimmed mean over the
differences of the log-transformed gene-count fold-change between the
sample and the reference.
\item "RLE", relative log expression, RLE uses a pseudo-reference calculated
using the geometric mean of the gene-specific abundances over all
samples. The scaling factors are then calculated as the median of the
gene counts ratios between the samples and the reference.
\item "CSS": cumulative sum scaling, calculates scaling factors as the
cumulative sum of gene abundances up to a data-derived threshold.
\item "CLR": centered log-ratio normalization.
}}

\item{norm_para}{named \code{list}. other arguments passed to specific
normalization methods.  Most users will not need to pass any additional
arguments here.}

\item{p_adjust}{method for multiple test correction, default \code{none},
for more details see \link[stats:p.adjust]{stats::p.adjust}.}

\item{alpha}{significance level for each of the statistical tests,
default 0.05.}

\item{theta}{lower bound for the proportion for the W-statistic, default 0.7.}

\item{test}{character, the test to dtermine the p value of log ratio,
one of "aov", "wilcox.test",  "kruskal.test".}

\item{...}{additional arguments passed to the test function.}
}
\value{
a \linkS4class{microbiomeMarker} object, in which the \code{slot} of \code{marker_table}
contains four variables:
\itemize{
\item \code{feature}, significantly different features.
\item \code{enrich_group}, the class of the differential features enriched.
\item \code{effect_size}, differential means for two groups, or F statistic for more
than two groups
\item \code{W}, the W-statistic, number of features that a single feature is tested to
be significantly different against.
}
}
\description{
Perform significant test by comparing the pairwise log ratios (based on
relative abundances) between all features.
}
\details{
The developers of this method recommend the following significance tests
if there are 2 groups, use non-parametric Wilcoxon rank sum test
\code{\link[stats:wilcox.test]{stats::wilcox.test()}}. If there are more than 2 groups, use nonparametric
\code{\link[stats:kruskal.test]{stats::kruskal.test()}} or one-way ANOVA \code{\link[stats:aov]{stats::aov()}}.
}
\references{
Mandal et al. "Analysis of composition of microbiomes: a novel
method for studying microbial composition", Microbial Ecology in Health
& Disease, (2015), 26.
}
\author{
Huang Lin, Yang Cao
}
