% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA-ancombc.R
\name{run_ancombc}
\alias{run_ancombc}
\title{Differential analysis of compositions of microbiomes with bias correction
(ANCOM-BC).}
\usage{
run_ancombc(
  ps,
  group,
  confounders = character(0),
  contrast = NULL,
  taxa_rank = "all",
  transform = c("identity", "log10", "log10p"),
  norm = "none",
  norm_para = list(),
  p_adjust = c("none", "fdr", "bonferroni", "holm", "hochberg", "hommel", "BH", "BY"),
  prv_cut = 0.1,
  lib_cut = 0,
  struc_zero = FALSE,
  neg_lb = FALSE,
  tol = 1e-05,
  max_iter = 100,
  conserve = FALSE,
  pvalue_cutoff = 0.05
)
}
\arguments{
\item{ps}{a \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}} object, which consists of a feature
table, a sample metadata and a taxonomy table.}

\item{group}{the name of the group variable in metadata. Specifying
\code{group} is required for detecting structural zeros and performing
global test.}

\item{confounders}{character vector, the confounding variables to be adjusted.
default \code{character(0)}, indicating no confounding variable.}

\item{contrast}{this parameter only used for two groups comparison while
there are multiple groups. For more please see the following details.}

\item{taxa_rank}{character to specify taxonomic rank to perform
differential analysis on. Should be one of
\code{phyloseq::rank_names(phyloseq)}, or "all" means to summarize the taxa by
the top taxa ranks (\code{summarize_taxa(ps, level = rank_names(ps)[1])}), or
"none" means perform differential analysis on the original taxa
(\code{taxa_names(phyloseq)}, e.g., OTU or ASV).}

\item{transform}{character, the methods used to transform the microbial
abundance. See \code{\link[=transform_abundances]{transform_abundances()}} for more details. The
options include:
\itemize{
\item "identity", return the original data without any transformation
(default).
\item "log10", the transformation is \code{log10(object)}, and if the data contains
zeros the transformation is \code{log10(1 + object)}.
\item "log10p", the transformation is \code{log10(1 + object)}.
}}

\item{norm}{the methods used to normalize the microbial abundance data. See
\code{\link[=normalize]{normalize()}} for more details.
Options include:
\itemize{
\item "none": do not normalize.
\item "rarefy": random subsampling counts to the smallest library size in the
data set.
\item "TSS": total sum scaling, also referred to as "relative abundance", the
abundances were normalized by dividing the corresponding sample library
size.
\item "TMM": trimmed mean of m-values. First, a sample is chosen as reference.
The scaling factor is then derived using a weighted trimmed mean over the
differences of the log-transformed gene-count fold-change between the
sample and the reference.
\item "RLE", relative log expression, RLE uses a pseudo-reference calculated
using the geometric mean of the gene-specific abundances over all
samples. The scaling factors are then calculated as the median of the
gene counts ratios between the samples and the reference.
\item "CSS": cumulative sum scaling, calculates scaling factors as the
cumulative sum of gene abundances up to a data-derived threshold.
\item "CLR": centered log-ratio normalization.
\item "CPM": pre-sample normalization of the sum of the values to 1e+06.
}}

\item{norm_para}{named \code{list}. other arguments passed to specific
normalization methods.  Most users will not need to pass any additional
arguments here.}

\item{p_adjust}{method to adjust p-values by. Default is "holm".
Options include "holm", "hochberg", "hommel", "bonferroni", "BH", "BY",
"fdr", "none". See \code{\link[stats:p.adjust]{stats::p.adjust()}} for more details.}

\item{prv_cut}{a numerical fraction between 0 and 1. Taxa with prevalences
less than \code{prv_cut} will be excluded in the analysis. Default
is 0.10.}

\item{lib_cut}{a numerical threshold for filtering samples based on library
sizes. Samples with library sizes less than \code{lib_cut} will be excluded
in the analysis. Default is 0, i.e. do not filter any sample.}

\item{struc_zero}{whether to detect structural zeros. Default is FALSE.}

\item{neg_lb}{whether to classify a taxon as a structural zero in the
corresponding study group using its asymptotic lower bound.
Default is FALSE.}

\item{tol}{the iteration convergence tolerance for the E-M algorithm.
Default is 1e-05.}

\item{max_iter}{the maximum number of iterations for the E-M algorithm.
Default is 100.}

\item{conserve}{whether to use a conservative variance estimate of
the test statistic. It is recommended if the sample size is small and/or
the number of differentially abundant taxa is believed to be large.
Default is FALSE.}

\item{pvalue_cutoff}{level of significance. Default is 0.05.}
}
\value{
a \code{\linkS4class{microbiomeMarker}} object.
}
\description{
Differential abundance analysis for microbial absolute abundance data. This
function is a wrapper of \code{\link[ANCOMBC:ancombc]{ANCOMBC::ancombc()}}.
}
\details{
\code{contrast} must be a two length character or \code{NULL} (default). It is only
required to set manually for two groups comparison when there are multiple
groups. The order determines the direction of comparison, the first element
is used to specify the reference group (control). This means that, the first
element is the denominator for the fold change, and the second element is
used as baseline (numerator for fold change). Otherwise, users do required
to concern this parameter (set as default \code{NULL}), and if there are
two groups, the first level of groups will set as the reference group; if
there are multiple groups, it will perform an ANOVA-like testing to find
markers which difference in any of the groups.
}
\examples{
data(enterotypes_arumugam)
ps <- phyloseq::subset_samples(
    enterotypes_arumugam,
    Enterotype \%in\% c("Enterotype 3", "Enterotype 2")
)
if (requireNamespace("microbiome", quietly = TRUE)) {
    run_ancombc(ps, group = "Enterotype")
} else {
    message("The 'mirobiome' package is not installed, please install it to use this example")
}
}
\references{
Lin, Huang, and Shyamal Das Peddada. "Analysis of compositions of microbiomes
with bias correction." Nature communications 11.1 (2020): 1-11.
}
\seealso{
\code{\link[ANCOMBC:ancombc]{ANCOMBC::ancombc}}
}
