% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-two-groups.R
\name{run_t_test}
\alias{run_t_test}
\title{run t test or welch test}
\usage{
run_t_test(abd_group, conf_level = 0.95, var_equal = FALSE)
}
\arguments{
\item{abd_group}{a two length list, each element represents the feature
abundance of a group}

\item{conf_level}{numeric, confidence level of the interval, default 0.95}

\item{var_equal}{a logical variable indicating whether to treat the two
variances as being equal. If TRUE then the pooled variance is used to
estimate the variance otherwise the Welch (or Satterthwaite) approximation
to the degrees of freedom is used.}
}
\description{
run t test or welch test
}
\seealso{
\code{\link[stats:t.test]{stats::t.test()}}
}
