% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-two-groups.R
\name{run_welch_test}
\alias{run_welch_test}
\title{welch test}
\usage{
run_welch_test(abd_group, conf_level = 0.95)
}
\arguments{
\item{abd_group}{a two length list, each element represents the feature
abundance of a group}

\item{conf_level}{numeric, confidence level of the interval, default 0.95}
}
\description{
welch test
}
