% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-cladogram.R
\name{set_marker_annotation}
\alias{set_marker_annotation}
\title{set legend for multiple geom_cladelabel layers}
\usage{
set_marker_annotation(p, color, label, size = 5, shape = 22, ...)
}
\arguments{
\item{p}{a ggtree object}

\item{color}{a color vector}

\item{label}{a character vector, with the same length with \code{color}}

\item{size}{size of shape, default \code{5}}

\item{shape}{shape of label, default \code{22}}

\item{...}{extra arguments passed to  \code{\link[ggplot2:guide_legend]{ggplot2::guide_legend()}},
e.g. \code{ncol}, more details see \code{\link[ggplot2:guide_legend]{ggplot2::guide_legend()}}.}
}
\value{
an updated \code{ggtree} object
}
\description{
This function can be used to set the microbiome marker annotations
}
\seealso{
\code{\link[ggplot2:guide_legend]{ggplot2::guide_legend()}}
}
