% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset-marker.R
\name{subset_marker}
\alias{subset_marker}
\title{Subset microbiome markers}
\usage{
subset_marker(mm, ...)
}
\arguments{
\item{mm}{a \code{\linkS4class{microbiomeMarker}} or \code{\linkS4class{marker_table}} object.}

\item{...}{the subsetting expression passed to \code{\link[base:subset]{base::subset()}}.}
}
\value{
a subset object in the same class with \code{mm}.
}
\description{
Subset markers based on an expression related to the columns and values
within the \code{marker_table} slot of \code{mm}.
}
\examples{
data(enterotypes_arumugam)
mm <- run_limma_voom(
    enterotypes_arumugam,
    "Enterotype",
    contrast = c("Enterotype 3", "Enterotype 2"),
    pvalue_cutoff = 0.01,
    p_adjust = "none"
)
subset_marker(mm, pvalue < 0.005)
}
