% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_qiime2.R
\name{subset_taxa_in_feature}
\alias{subset_taxa_in_feature}
\title{Subset taxa according to the taxa in feature table}
\usage{
subset_taxa_in_feature(taxa_tab, feature_tab)
}
\description{
Subset taxa according to the taxa in feature table
}
\keyword{internal}
