% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_taxa.R
\name{summarize_taxa}
\alias{summarize_taxa}
\title{Summarize taxa into a taxonomic level within each sample}
\usage{
summarize_taxa(ps, level = "Kingdom", absolute = FALSE, sep = "|")
}
\arguments{
\item{ps}{a \code{\link[phyloseq]{phyloseq-class}} object.}

\item{level}{taxonomic level to summarize, default \code{Kingdom}.}

\item{absolute}{logical, whether return the absolute abundance or
relative abundance, default \code{FALSE}
FALSE.}

\item{sep}{a character string to separate the taxonomic levels.}
}
\value{
a data frame, each row represents a taxa, where each col represents
the taxa abundance of each sample.
}
\description{
Provids summary information of the representation of a taxonomic levels within
each sample.
}
