% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA-comparing.R
\name{summary.compareDA}
\alias{summary.compareDA}
\title{Summary differential analysis methods comparison results}
\usage{
\method{summary}{compareDA}(
  object,
  sort = c("score", "auc", "fpr", "power"),
  boot = TRUE,
  boot_n = 1000L,
  prob = c(0.05, 0.95),
  ...
)
}
\arguments{
\item{object}{an \code{compareDA} object, output from \code{\link[=compare_DA]{compare_DA()}}.}

\item{sort}{character string specifying sort method. Possibilities are
"score" which is calculated as \eqn{(auc - 0.5) * power - fdr}, "auc" for
area under the ROC curve, "fpr" for false positive rate, "power" for
empirical power.}

\item{boot}{logical, whether use bootstrap for confidence limites of the
score, default \code{TRUE}. Recommended to be \code{TRUE} unless \code{n_rep} is larger
then 100 in \code{\link[=compare_DA]{compare_DA()}}.}

\item{boot_n}{integer, number of bootstraps, default 1000L.}

\item{prob}{two length numeric vector, confidence limits for score, default
\code{c(0.05, 0.95)}.}

\item{...}{extra arguments affecting the summary produced.}
}
\value{
a \code{data.frame} containing measurements for differential analysis
methods:
\itemize{
\item \code{call}: differential analysis commands.
\item \code{auc}: area under curve of ROC.
\item \code{fpr}: false positive rate
\item \code{power}: empirical power.
\item \code{fdr}: false discover7y rate.
\item \code{score}: score whch is calculated as \eqn{(auc - 0.5) * power - fdr}.
\item \verb{score_*}: confidence limits of score.
}
}
\description{
Summary differential analysis methods comparison results
}
