% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-multiple-groups.R
\name{test_multiple_groups}
\alias{test_multiple_groups}
\title{Statistical test for multiple groups}
\usage{
test_multiple_groups(
  ps,
  group,
  rank_name,
  method = c("anova", "kruskal"),
  p_adjust = c("none", "fdr", "bonferroni", "holm", "hochberg", "hommel", "BH", "BY"),
  p_value_cutoff = 0.05,
  effect_size_cutoff = NULL
)
}
\arguments{
\item{ps}{a \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}} object}

\item{group}{character, the variable to set the group}

\item{rank_name}{character, taxonomic names of \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}}
to compare}

\item{method}{test method, must be one of "anova" or "kruskal"}

\item{p_adjust}{method for multiple test correction, default \code{none},
for more details see \link[stats:p.adjust]{stats::p.adjust}.}

\item{p_value_cutoff}{numeric, p value cutoff, default 0.05}

\item{effect_size_cutoff}{numeric, cutoff of effect size default \code{NULL}
which means no effect size filter}
}
\description{
Statistical test for multiple groups
}
\seealso{
\code{\link[=posthoc_test]{posthoc_test()}}
}
