% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-two-groups.R
\name{test_two_groups}
\alias{test_two_groups}
\title{Statistical test between two groups}
\usage{
test_two_groups(
  ps,
  groups,
  rank_name,
  method = "welch.test",
  p_adjust_method = "fdr",
  conf_level = 0.95
)
}
\arguments{
\item{ps}{a \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}} object}

\item{groups}{character, the variable to set the group}

\item{rank_name}{character, taxonomic names of \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}}
to compare}

\item{method}{test method}

\item{p_adjust_method}{method for multiple test correction}

\item{conf_level}{numeric, confidence level of interval}
}
\description{
Statistical test between two groups
}
\author{
Yang Cao
}
