% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-effect-size.R
\name{plot_ef_bar}
\alias{plot_ef_bar}
\alias{ef-barplot,ef-dotplot}
\alias{plot_ef_dot}
\title{bar and dot plot of effect size of microbiomeMarker data}
\usage{
plot_ef_bar(mm, label_level = 1, max_label_len = 60, markers = NULL)

plot_ef_dot(mm, label_level = 1, max_label_len = 60, markers = NULL)
}
\arguments{
\item{mm}{a \code{\linkS4class{microbiomeMarker}} object}

\item{label_level}{integer, number of label levels to be displayed, default
\code{1}, \code{0} means display the full name of the feature}

\item{max_label_len}{integer, maximum number of characters of feature label,
default \code{60}}

\item{markers}{character vector, markers to display, default \code{NULL},
indicating plot all markers.}
}
\value{
a ggplot project
}
\description{
bar and dot plot of effect size microbiomeMarker data. This function returns
a \code{ggplot2} object that can be saved or further customized using \strong{ggplot2}
package.
}
\examples{
data(enterotypes_arumugam)
mm <- run_limma_voom(
    enterotypes_arumugam,
    "Enterotype",
    contrast = c("Enterotype 3", "Enterotype 2"),
    pvalue_cutoff = 0.01,
    p_adjust = "none"
)
plot_ef_bar(mm)
}
