% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-methods.R
\name{[}
\alias{[}
\alias{[,marker_table,ANY,ANY,ANY-method}
\title{Extract \code{marker_table} object}
\usage{
\S4method{[}{marker_table,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{a \code{\link{marker_table}} object.}

\item{i, j}{elements to extract.}

\item{...}{see \code{\link[base:Extract]{base::Extract()}}.}

\item{drop}{ignored now.}
}
\value{
a \code{marker_table} object.
}
\description{
Operators acting on \code{marker_table} to extract parts.
}
\seealso{
\code{\link[base:Extract]{base::Extract()}}
}
