% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lefse.R
\name{lefse}
\alias{lefse}
\title{Liner discriminant analysis (LDA) effect size (LEFSe) analysis}
\usage{
lefse(ps, tax_rank, class, p_cutoff = 0.05, lda_cutoff = 2)
}
\arguments{
\item{ps}{a \code{\link[phyloseq]{phyloseq-class}} object}

\item{tax_rank}{character, rank level}

\item{class}{character, the column name to specify the class}

\item{p_cutoff}{numeric, p value cutoff, default 0.05}

\item{lda_cutoff}{numeric, lda score cutoff, default 2}
}
\description{
Perform Metagenomic LEFSe analysis based on phyloseq object.
}
\references{
Segata, Nicola, et al. Metagenomic biomarker discovery and
explanation. Genome biology 12.6 (2011): R60.
}
\author{
Yang Cao \email{yiluheihei@gmail.com}
}
