% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllGenerics.R
\docType{class}
\name{microbiomeMarker-class}
\alias{microbiomeMarker-class}
\alias{show,microbiomeMarker-method}
\title{The main class for microbiomeMarker data}
\usage{
\S4method{show}{microbiomeMarker}(object)
}
\arguments{
\item{object}{a \code{microbiomeMarker-class} object}
}
\value{
a \code{\linkS4class{microbiomeMarker}} object.
}
\description{
\code{microbiomeMarker-class} is inherited from the \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}}
by adding a custom slot \code{microbiome_marker} to save the differential analysis
results. And it provides a seamless interface with \strong{phyloseq}, which makes
\strong{microbiomeMarker} simple and easy to use. For more details on see the
document of \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}}.
}
\section{Slots}{

\describe{
\item{\code{marker_table}}{a data.frame, a \code{\linkS4class{marker_table}} object.}

\item{\code{norm_method}}{character, method used to normalize the input \code{phyloseq}
object.}

\item{\code{diff_method}}{character, method used for marker identification.}
}}

\seealso{
\code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}}, \code{\linkS4class{marker_table}},
\code{\link[=summarize_taxa]{summarize_taxa()}}
}
