% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiomeMarker.R
\docType{package}
\name{microbiomeMarker}
\alias{microbiomeMarker}
\title{microbiomeMarker: A package for micribiome biomarker discovery}
\description{
The microboimeMarker package provides several methods to identify micribome
biomarker, such as lefse and statistical analysis.
}
