% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{microbiomeMarker}
\alias{microbiomeMarker}
\title{Build microbiomeMarker-class objects}
\usage{
microbiomeMarker(
  marker_table = NULL,
  norm_method = NULL,
  diff_method = NULL,
  ...
)
}
\arguments{
\item{marker_table}{a \code{\linkS4class{marker_table}} object differtial analysis.}

\item{norm_method}{character, method used to normalize the input \code{phyloseq}
object.}

\item{diff_method}{character, method used for microbiome marker
identification.}

\item{...}{arguments passed to \code{\link[phyloseq:phyloseq]{phyloseq::phyloseq()}}}
}
\value{
a \code{\linkS4class{microbiomeMarker}} object.
}
\description{
This the constructor to build the \code{\linkS4class{microbiomeMarker}} object, don't
use the \code{new()} constructor.
}
\examples{
microbiomeMarker(
    marker_table = marker_table(data.frame(
        feature = c("speciesA", "speciesB"),
        enrich_group = c("groupA", "groupB"),
        ef_logFC = c(-2, 2),
        pvalue = c(0.01, 0.01),
        padj = c(0.01, 0.01),
        row.names = c("marker1", "marker2")
    )),
    norm_method = "TSS",
    diff_method = "DESeq2",
    otu_table = otu_table(matrix(
        c(4, 1, 1, 4),
        nrow = 2, byrow = TRUE,
        dimnames = list(c("speciesA", "speciesB"), c("sample1", "sample2"))
    ),
    taxa_are_rows = TRUE
    ),
    tax_table = tax_table(matrix(
        c("speciesA", "speciesB"),
        nrow = 2,
        dimnames = list(c("speciesA", "speciesB"), "Species")
    )),
    sam_data = sample_data(data.frame(
        group = c("groupA", "groupB"),
        row.names = c("sample1", "sample2")
    ))
)
}
\seealso{
\code{\link[phyloseq:phyloseq]{phyloseq::phyloseq()}}
}
