% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-postHocTest.R
\name{plot_postHocTest}
\alias{plot_postHocTest}
\title{\code{postHocTest} plot}
\usage{
plot_postHocTest(pht, feature, step_increase = 0.12)
}
\arguments{
\item{pht}{a \code{\linkS4class{postHocTest}} object}

\item{feature}{character, to plot the post-toc test result of this feature}

\item{step_increase}{numeric vector with the increase in fraction of total
height for every additional comparison to minimize overlap, default \code{0.12}.}
}
\value{
a \code{ggplot} object
}
\description{
Visualize the result of post-hoc test using ggplot2
}
\examples{
data(enterotypes_arumugam)
ps <- phyloseq::subset_samples(
    enterotypes_arumugam,
    Enterotype \%in\% c("Enterotype 3", "Enterotype 2", "Enterotype 1")
) \%>\%
    phyloseq::subset_taxa(Phylum == "Bacteroidetes")
pht <- run_posthoc_test(ps, group = "Enterotype")
plot_postHocTest(pht, feature = "p__Bacteroidetes|g__Alistipes")
}
