% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-sl-roc.R
\name{plot_sl_roc}
\alias{plot_sl_roc}
\title{ROC curve of microbiome marker from supervised learning methods}
\usage{
plot_sl_roc(mm, group, nfolds = 3, nrepeats = 3, tune_length = 5, ...)
}
\arguments{
\item{mm}{a \linkS4class{microbiomeMarker} object.}

\item{group, nfolds, nrepeats, tune_length, ...}{same with the \code{run_sl()}.}
}
\value{
a \code{\link[ggplot2:ggplot]{ggplot2::ggplot}} object.
}
\description{
Show the ROC curve of the microbiome marker calculated by \code{run_sl}.
}
\examples{
data(enterotypes_arumugam)
# small example phyloseq object for test
ps_s <- phyloseq::subset_taxa(
    enterotypes_arumugam,
    Phylum \%in\% c("Firmicutes", "Bacteroidetes")
)

set.seed(2021)
mm <- run_sl(
    ps_s,
    group = "Gender",
    taxa_rank = "Genus",
    nfolds = 2,
    nrepeats = 1,
    top_n = 15,
    norm = "TSS",
    method = "LR",
)
plot_sl_roc(mm, group = "Gender")
}
\seealso{
\code{\link[=run_sl]{run_sl()}}
}
