% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA-ancom.R
\name{run_ancom}
\alias{run_ancom}
\title{Perform differential analysis using ANCOM}
\usage{
run_ancom(
  ps,
  group,
  confounders = character(0),
  taxa_rank = "all",
  transform = c("identity", "log10", "log10p"),
  norm = "TSS",
  norm_para = list(),
  p_adjust = c("none", "fdr", "bonferroni", "holm", "hochberg", "hommel", "BH", "BY"),
  pvalue_cutoff = 0.05,
  W_cutoff = 0.75
)
}
\arguments{
\item{ps}{a \code{\link[phyloseq]{phyloseq-class}} object.}

\item{group}{character, the variable to set the group.}

\item{confounders}{character vector, the confounding variables to be adjusted.
default \code{character(0)}, indicating no confounding variable.}

\item{taxa_rank}{character to specify taxonomic rank to perform
differential analysis on. Should be one of
\code{phyloseq::rank_names(phyloseq)}, or "all" means to summarize the taxa by
the top taxa ranks (\code{summarize_taxa(ps, level = rank_names(ps)[1])}), or
"none" means perform differential analysis on the original taxa
(\code{taxa_names(phyloseq)}, e.g., OTU or ASV).}

\item{transform}{character, the methods used to transform the microbial
abundance. See \code{\link[=transform_abundances]{transform_abundances()}} for more details. The
options include:
\itemize{
\item "identity", return the original data without any transformation.
\item "log10", the transformation is \code{log10(object)}, and if the data contains
zeros the transformation is \code{log10(1 + object)}.
\item "log10p", the transformation is \code{log10(1 + object)}.
}}

\item{norm}{the methods used to normalize the microbial abundance data. See
\code{\link[=normalize]{normalize()}} for more details.
Options include:
\itemize{
\item "none": do not normalize.
\item "rarefy": random subsampling counts to the smallest library size in the
data set.
\item "TSS": total sum scaling, also referred to as "relative abundance", the
abundances were normalized by dividing the corresponding sample library
size.
\item "TMM": trimmed mean of m-values. First, a sample is chosen as reference.
The scaling factor is then derived using a weighted trimmed mean over
the differences of the log-transformed gene-count fold-change between
the sample and the reference.
\item "RLE", relative log expression, RLE uses a pseudo-reference calculated
using the geometric mean of the gene-specific abundances over all
samples. The scaling factors are then calculated as the median of the
gene counts ratios between the samples and the reference.
\item "CSS": cumulative sum scaling, calculates scaling factors as the
cumulative sum of gene abundances up to a data-derived threshold.
\item "CLR": centered log-ratio normalization.
\item "CPM": pre-sample normalization of the sum of the values to 1e+06.
}}

\item{norm_para}{named \code{list}. other arguments passed to specific
normalization methods.  Most users will not need to pass any additional
arguments here.}

\item{p_adjust}{method for multiple test correction, default \code{none},
for more details see \link[stats:p.adjust]{stats::p.adjust}.}

\item{pvalue_cutoff}{significance level for each of the statistical tests,
default 0.05.}

\item{W_cutoff}{lower bound for the proportion for the W-statistic, default
0.7.}
}
\value{
a \linkS4class{microbiomeMarker} object, in which the \code{slot} of
\code{marker_table} contains four variables:
\itemize{
\item \code{feature}, significantly different features.
\item \code{enrich_group}, the class of the differential features enriched.
\item \code{effect_size}, differential means for two groups, or F statistic for more
than two groups.
\item \code{W}, the W-statistic, number of features that a single feature is tested
to be significantly different against.
}
}
\description{
Perform significant test by comparing the pairwise log ratios between all
features.
}
\details{
In an experiment with only two treatments, this tests the following
hypothesis for feature \eqn{i}:

\deqn{H_{0i}: E(log(\mu_i^1)) =  E(log(\mu_i^2))}

where \eqn{\mu_i^1} and \eqn{\mu_i^2} are the mean abundances for feature
\eqn{i} in the two groups.

The developers of this method recommend the following significance tests
if there are 2 groups, use non-parametric Wilcoxon rank sum test
\code{\link[stats:wilcox.test]{stats::wilcox.test()}}. If there are more than 2 groups, use nonparametric
\code{\link[stats:kruskal.test]{stats::kruskal.test()}} or one-way ANOVA \code{\link[stats:aov]{stats::aov()}}.
}
\examples{
\donttest{
data(enterotypes_arumugam)
ps <- phyloseq::subset_samples(
    enterotypes_arumugam,
    Enterotype \%in\% c("Enterotype 3", "Enterotype 2")
)
run_ancom(ps, group = "Enterotype")
}
}
\references{
Mandal et al. "Analysis of composition of microbiomes: a novel
method for studying microbial composition", Microbial Ecology in Health
& Disease, (2015), 26.
}
\author{
Huang Lin, Yang Cao
}
