% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancombc.R
\name{run_ancombc}
\alias{run_ancombc}
\title{Differential analysis of compositions of microbiomes with bias correction
(ANCOM-BC).}
\usage{
run_ancombc(
  ps,
  formula,
  taxa_rank = "all",
  transform = c("identity", "log10", "log10p"),
  norm = "none",
  norm_para = list(),
  p_adjust = "holm",
  zero_cut = 0.9,
  lib_cut = 0,
  group_var = NULL,
  struc_zero = FALSE,
  neg_lb = FALSE,
  tol = 1e-05,
  max_iter = 100,
  conserve = FALSE,
  alpha = 0.05
)
}
\arguments{
\item{ps}{a \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}} object, which consists of a feature
table, a sample metadata and a taxonomy table.}

\item{formula}{the character string expresses how the microbial absolute
abundances for each taxon depend on the variables in metadata.}

\item{taxa_rank}{character to specify taxonomic rank to perform
differential analysis on. Should be one of \code{phyloseq::rank_names(phyloseq)},
or "all" means to summarize the taxa by the top taxa ranks
(\code{summarize_taxa(ps, level = rank_names(ps)[1])}), or "none" means perform
differential analysis on the original taxa (\code{taxa_names(phyloseq)}, e.g.,
OTU or ASV).}

\item{transform}{character, the methods used to transform the microbial
abundance. See \code{\link[=transform_abundances]{transform_abundances()}} for more details. The
options include:
\itemize{
\item "identity", return the original data without any transformation (default).
\item "log10", the transformation is \code{log10(object)}, and if the data contains
zeros the transformation is \code{log10(1 + object)}.
\item "log10p", the transformation is \code{log10(1 + object)}.
}}

\item{norm}{the methods used to normalize the microbial abundance data. See
\code{\link[=normalize]{normalize()}} for more details.
Options include:
\itemize{
\item "none": do not normalize.
\item "rarefy": random subsampling counts to the smallest library size in the
data set.
\item "TSS": total sum scaling, also referred to as "relative abundance", the
abundances were normalized by dividing the corresponding sample library
size.
\item "TMM": trimmed mean of m-values. First, a sample is chosen as reference.
The scaling factor is then derived using a weighted trimmed mean over the
differences of the log-transformed gene-count fold-change between the
sample and the reference.
\item "RLE", relative log expression, RLE uses a pseudo-reference calculated
using the geometric mean of the gene-specific abundances over all
samples. The scaling factors are then calculated as the median of the
gene counts ratios between the samples and the reference.
\item "CSS": cumulative sum scaling, calculates scaling factors as the
cumulative sum of gene abundances up to a data-derived threshold.
\item "CLR": centered log-ratio normalization.
\item "CPM": pre-sample normalization of the sum of the values to 1e+06.
}}

\item{norm_para}{named \code{list}. other arguments passed to specific
normalization methods.  Most users will not need to pass any additional
arguments here.}

\item{p_adjust}{method to adjust p-values by. Default is "holm".
Options include "holm", "hochberg", "hommel", "bonferroni", "BH", "BY",
"fdr", "none". See \code{\link[stats:p.adjust]{stats::p.adjust()}} for more details.}

\item{zero_cut}{a numerical fraction between 0 and 1. Taxa with proportion of
zeroes greater than \code{zero_cut} will be excluded in the analysis. Default
is 0.90.}

\item{lib_cut}{a numerical threshold for filtering samples based on library
sizes. Samples with library sizes less than \code{lib_cut} will be excluded
in the analysis. Default is 0, i.e. do not filter any sample.}

\item{group_var}{the name of the group variable in metadata. Specifying
\code{group_var} is required for detecting structural zeros and performing
global test.}

\item{struc_zero}{whether to detect structural zeros. Default is FALSE.}

\item{neg_lb}{whether to classify a taxon as a structural zero in the
corresponding study group using its asymptotic lower bound.
Default is FALSE.}

\item{tol}{the iteration convergence tolerance for the E-M algorithm.
Default is 1e-05.}

\item{max_iter}{the maximum number of iterations for the E-M algorithm.
Default is 100.}

\item{conserve}{whether to use a conservative variance estimate of
the test statistic. It is recommended if the sample size is small and/or
the number of differentially abundant taxa is believed to be large.
Default is FALSE.}

\item{alpha}{level of significance. Default is 0.05.}
}
\description{
Differential abundance analysis for microbial absolute abundance data. This
function is a wrapper of \code{\link[ANCOMBC:ancombc]{ANCOMBC::ancombc()}}.
}
\references{
Lin, Huang, and Shyamal Das Peddada. "Analysis of compositions of microbiomes
with bias correction." Nature communications 11.1 (2020): 1-11.
}
\seealso{
\code{\link[ANCOMBC:ancombc]{ANCOMBC::ancombc}}
}
