% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2.R
\name{run_deseq2}
\alias{run_deseq2}
\title{Perform DESeq analysis}
\usage{
run_deseq2(
  ps,
  group_var,
  subgroup1,
  subgroup2,
  transform = c("identity", "log10", "log10p"),
  test = c("Wald", "LRT"),
  fitType = c("parametric", "local", "mean", "glmGamPoi"),
  sfType = "poscounts",
  betaPrior = FALSE,
  modelMatrixType,
  useT = FALSE,
  minmu = if (fitType == "glmGamPoi") 1e-06 else 0.5,
  p_adjust = c("none", "fdr", "bonferroni", "holm", "hochberg", "hommel", "BH", "BY"),
  p_value_cutoff = 0.05,
  ...
)
}
\arguments{
\item{ps}{ps a \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}} object.}

\item{group_var}{character, the variable to set the group, must be one of
the var of the sample metadata}

\item{subgroup1, subgroup2}{character, subgroups to be compared.}

\item{transform}{character, the methods used to transform the microbial
abundance. See \code{\link[=transform_abundances]{transform_abundances()}} for more details. The
options include:
\itemize{
\item "identity", return the original data without any transformation (default).
\item "log10", the transformation is \code{log10(object)}, and if the data contains
zeros the transformation is \code{log10(1 + object)}.
\item "log10p", the transformation is \code{log10(1 + object)}.
}}

\item{test, fitType, sfType, betaPrior, modelMatrixType, useT, minmu}{these seven
parameters are inherited form \code{\link[DESeq2:DESeq]{DESeq2::DESeq()}}.
\itemize{
\item \code{test}, should be either "Wald" or "LRT", which will
then use either Wald significance tests, or the likelihood ratio test on
the difference in deviance between a full and reduced model formula.
\item \code{fitType}, either "parametric", "local", "mean", or "glmGamPoi" for the
type of fitting of dispersions to the mean intensity.
\item \code{sfType}, either "ratio", "poscounts", or "iterate" for the type of size
factor estimation.
\item \code{betaPrior}, whether or not to put a zero-mean normal prior on the
non-intercept coefficients.
\item \code{modelMatrixType}, either "standard" or "expanded", which describe how
the model matrix,
\item \code{useT}, logical, where Wald statistics are assumed to follow a standard
Normal.
\item \code{minmu}, lower bound on the estimated count for fitting gene-wise
dispersion.
}

For more details, see \code{\link[DESeq2:DESeq]{DESeq2::DESeq()}}.}

\item{p_adjust}{method for multiple test correction, default \code{none}, for
more details see \link[stats:p.adjust]{stats::p.adjust}.}

\item{p_value_cutoff}{p_value_cutoff numeric, p value cutoff, default 0.05.}

\item{...}{extra parameters passed to \code{\link[DESeq2:DESeq]{DESeq2::DESeq()}}.}
}
\value{
a \code{\linkS4class{microbiomeMarker}} object.
}
\description{
Differential expression analysis based on the Negative Binomial distribution
using \strong{DESeq2}.
}
\seealso{
\code{\link[DESeq2:results]{DESeq2::results()}},\code{\link[DESeq2:DESeq]{DESeq2::DESeq()}}
}
