% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA-metagenomeSeq.R
\name{run_metagenomeseq}
\alias{run_metagenomeseq}
\title{metagenomeSeq differential analysis}
\usage{
run_metagenomeseq(
  ps,
  group,
  confounders = character(0),
  contrast = NULL,
  taxa_rank = "all",
  transform = c("identity", "log10", "log10p"),
  norm = "CSS",
  norm_para = list(),
  method = c("ZILN", "ZIG"),
  p_adjust = c("none", "fdr", "bonferroni", "holm", "hochberg", "hommel", "BH", "BY"),
  pvalue_cutoff = 0.05,
  ...
)
}
\arguments{
\item{ps}{ps a \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}} object.}

\item{group}{character, the variable to set the group, must be one of
the var of the sample metadata.}

\item{confounders}{character vector, the confounding variables to be adjusted.
default \code{character(0)}, indicating no confounding variable.}

\item{contrast}{this parameter only used for two groups comparison while
there are multiple groups. For more please see the following details.}

\item{taxa_rank}{character to specify taxonomic rank to perform
differential analysis on. Should be one of \code{phyloseq::rank_names(ps)},
or "all" means to summarize the taxa by the top taxa ranks
(\code{summarize_taxa(ps, level = rank_names(ps)[1])}), or "none" means perform
differential analysis on the original taxa (\code{taxa_names(ps)}, e.g.,
OTU or ASV).}

\item{transform}{character, the methods used to transform the microbial
abundance. See \code{\link[=transform_abundances]{transform_abundances()}} for more details. The
options include:
\itemize{
\item "identity", return the original data without any transformation
(default).
\item "log10", the transformation is \code{log10(object)}, and if the data contains
zeros the transformation is \code{log10(1 + object)}.
\item "log10p", the transformation is \code{log10(1 + object)}.
}}

\item{norm}{the methods used to normalize the microbial abundance data. See
\code{\link[=normalize]{normalize()}} for more details.
Options include:
\itemize{
\item "none": do not normalize.
\item "rarefy": random subsampling counts to the smallest library size in the
data set.
\item "TSS": total sum scaling, also referred to as "relative abundance", the
abundances were normalized by dividing the corresponding sample library
size.
\item "TMM": trimmed mean of m-values. First, a sample is chosen as reference.
The scaling factor is then derived using a weighted trimmed mean over the
differences of the log-transformed gene-count fold-change between the
sample and the reference.
\item "RLE", relative log expression, RLE uses a pseudo-reference calculated
using the geometric mean of the gene-specific abundances over all
samples. The scaling factors are then calculated as the median of the
gene counts ratios between the samples and the reference.
\item "CSS": cumulative sum scaling, calculates scaling factors as the
cumulative sum of gene abundances up to a data-derived threshold.
\item "CLR": centered log-ratio normalization.
\item "CPM": pre-sample normalization of the sum of the values to 1e+06.
}}

\item{norm_para}{arguments passed to specific normalization methods.}

\item{method}{character, which model used for differential analysis,
"ZILN" (Zero-inflated Log-Normal mixture model)" or "ZIG" (Zero-inflated
Gaussian mixture model). And the zero-inflated log-normal model is
preferred due to the high sensitivity and low FDR.}

\item{p_adjust}{method for multiple test correction, default \code{none},
for more details see \link[stats:p.adjust]{stats::p.adjust}.}

\item{pvalue_cutoff}{numeric, p value cutoff, default 0.05}

\item{...}{extra arguments passed to the model. more details see
\code{\link[metagenomeSeq:fitFeatureModel]{metagenomeSeq::fitFeatureModel()}} and \code{\link[metagenomeSeq:fitZig]{metagenomeSeq::fitZig()}},
e.g. \code{control} (can be setted using \code{\link[metagenomeSeq:zigControl]{metagenomeSeq::zigControl()}}) for
\code{\link[metagenomeSeq:fitZig]{metagenomeSeq::fitZig()}}.}
}
\value{
a \code{\linkS4class{microbiomeMarker}} object.
}
\description{
Differential expression analysis based on the Zero-inflated Log-Normal
mixture model or Zero-inflated Gaussian mixture model using metagenomeSeq.
}
\details{
metagnomeSeq provides two differential analysis methods, zero-inflated
log-normal mixture model (implemented in
\code{\link[metagenomeSeq:fitFeatureModel]{metagenomeSeq::fitFeatureModel()}}) and zero-inflated Gaussian mixture
model (implemented in \code{\link[metagenomeSeq:fitZig]{metagenomeSeq::fitZig()}}). We recommend
fitFeatureModel over fitZig due to high sensitivity and low FDR. Both
\code{\link[metagenomeSeq:fitFeatureModel]{metagenomeSeq::fitFeatureModel()}} and \code{\link[metagenomeSeq:fitZig]{metagenomeSeq::fitZig()}} require
the abundance profiles before normalization.

For \code{\link[metagenomeSeq:fitZig]{metagenomeSeq::fitZig()}}, the output column is the coefficient of
interest, and logFC column in the output of
\code{\link[metagenomeSeq:fitFeatureModel]{metagenomeSeq::fitFeatureModel()}} is analogous to coefficient. Thus,
logFC is really just the estimate the coefficient of interest in
\code{\link[metagenomeSeq:fitFeatureModel]{metagenomeSeq::fitFeatureModel()}}. For more details see
these question \href{https://support.bioconductor.org/p/94138/}{Difference between fitFeatureModel and fitZIG in metagenomeSeq}.

\code{contrast} must be a two length character or \code{NULL} (default). It is only
required to set manually for two groups comparison when there are multiple
groups. The order determines the direction of comparison, the first element
is used to specify the reference group (control). This means that, the first
element is the denominator for the fold change, and the second element is
used as baseline (numerator for fold change). Otherwise, users do required
to concern this paramerter (set as default \code{NULL}), and if there are
two groups, the first level of groups will set as the reference group; if
there are multiple groups, it will perform an ANOVA-like testing to find
markers which difference in any of the groups.

Of note, \code{\link[metagenomeSeq:fitFeatureModel]{metagenomeSeq::fitFeatureModel()}} is not allows for multiple
groups comparison.
}
\examples{
data(enterotypes_arumugam)
ps <- phyloseq::subset_samples(
    enterotypes_arumugam,
    Enterotype \%in\% c("Enterotype 3", "Enterotype 2")
)
run_metagenomeseq(ps, group = "Enterotype")
}
\references{
Paulson, Joseph N., et al. "Differential abundance analysis for microbial
marker-gene surveys." Nature methods 10.12 (2013): 1200-1202.
}
\author{
Yang Cao
}
