% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize-taxa.R
\name{summarize_taxa}
\alias{summarize_taxa}
\title{Summarize taxa into a taxonomic level within each sample}
\usage{
summarize_taxa(ps, level = rank_names(ps)[1], absolute = TRUE, sep = "|")
}
\arguments{
\item{ps}{a \code{\link[phyloseq]{phyloseq-class}} object.}

\item{level}{taxonomic level to summarize, default the top level rank of the
\code{ps}.}

\item{absolute}{logical, whether return the absolute abundance or
relative abundance, default \code{FALSE}.}

\item{sep}{a character string to separate the taxonomic levels.}
}
\value{
a \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}} object, where each row represents a
taxa, and each col represents the taxa abundance of each sample.
}
\description{
Provides summary information of the representation of a taxonomic levels
within each sample.
}
\examples{
data(enterotypes_arumugam)
summarize_taxa(enterotypes_arumugam)
}
