% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-two-groups.R
\name{test_two_groups}
\alias{test_two_groups}
\title{Statistical test between two groups}
\usage{
test_two_groups(
  ps,
  group,
  method = c("welch.test", "t.test", "white.test"),
  p_adjust = c("none", "fdr", "bonferroni", "holm", "hochberg", "hommel", "BH", "BY"),
  pvalue_cutoff = 0.05,
  diff_mean_cutoff = NULL,
  ratio_proportion_cutoff = NULL,
  conf_level = 0.95,
  nperm = 1000,
  ...
)
}
\arguments{
\item{ps}{a \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}} object}

\item{group}{character, the variable to set the group}

\item{method}{test method, must be one of "welch.test", "t.test" or
"white.test"}

\item{p_adjust}{method for multiple test correction, default \code{none},
for more details see \link[stats:p.adjust]{stats::p.adjust}.}

\item{pvalue_cutoff}{numeric, p value cutoff, default 0.05}

\item{diff_mean_cutoff, ratio_proportion_cutoff}{cutoff of different mean
proportions and ratio proportions, default \code{NULL} which means no effect size
filter.}

\item{conf_level}{numeric, confidence level of interval,}

\item{nperm}{interger, number of permutations for white non parametric t test
estimation}

\item{...}{extra arguments passed to \code{\link[=t.test]{t.test()}} or \code{\link[=fisher.test]{fisher.test()}}}
}
\value{
a \code{\linkS4class{microbiomeMarker}} object.
}
\description{
Statistical test between two groups
}
\author{
Yang Cao
}
