% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-two-groups.R
\name{test_two_groups}
\alias{test_two_groups}
\title{Statistical test between two groups}
\usage{
test_two_groups(
  ps,
  groups,
  rank_name,
  method = c("welch.test", "t.test"),
  p_adjust_method = c("none", "fdr", "bonferroni", "holm", "hochberg", "hommel", "BH",
    "BY"),
  p_value_cutoff = 0.05,
  diff_mean_cutoff = NULL,
  ratio_proportion_cutoff = NULL,
  conf_level = 0.95
)
}
\arguments{
\item{ps}{a \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}} object}

\item{groups}{character, the variable to set the group}

\item{rank_name}{character, taxonomic names of \code{\link[phyloseq:phyloseq-class]{phyloseq::phyloseq}}
to compare}

\item{method}{test method}

\item{p_adjust_method}{method for multiple test correction, default \code{none},
for more details see \link[stats:p.adjust]{stats::p.adjust}.}

\item{p_value_cutoff}{numeric, p value cutoff, default 0.05}

\item{diff_mean_cutoff, ratio_proportion_cutoff}{cutoff of different mean
proportions and ratio proportions, default \code{NULL} which means no effect size
filter.}

\item{conf_level}{numeric, confidence level of interval}
}
\description{
Statistical test between two groups
}
\author{
Yang Cao
}
